+++ src/realm/column.cpp
@@ -139,13 +139,16 @@ struct GetSizeFromRef {
     const ref_type m_ref;
     Allocator& m_alloc;
     size_t m_size;
-    GetSizeFromRef(ref_type r, Allocator& a): m_ref(r), m_alloc(a), m_size(0) {}
+    bool m_nullable;
+    GetSizeFromRef(ref_type r, Allocator& a, bool nullable): m_ref(r), m_alloc(a), m_size(0), m_nullable(nullable) {}
+
     template<class Col>
     void call() noexcept
     {
         m_size = Col::get_size_from_ref(m_ref, m_alloc);
     }
 };
+
 
 template<class Op>
 void col_type_deleg(Op& op, ColumnType type)
@@ -155,10 +158,14 @@ void col_type_deleg(Op& op, ColumnType type)
         case col_type_Bool:
         case col_type_OldDateTime:
         case col_type_Link:
-            op.template call<IntegerColumn>();
+            if(op.m_nullable)
+                op.template call<IntNullColumn>();
+            else
+                op.template call<IntegerColumn>();
             return;
         case col_type_Timestamp:
             op.template call<TimestampColumn>();
+            return;
         case col_type_String:
             op.template call<StringColumn>();
             return;
@@ -180,9 +187,13 @@ void col_type_deleg(Op& op, ColumnType type)
         case col_type_Double:
             op.template call<DoubleColumn>();
             return;
-        case col_type_Reserved4:
         case col_type_LinkList:
+            op.template call<LinkListColumn>();
+            return;
         case col_type_BackLink:
+            op.template call<BacklinkColumn>();
+            return;
+        case col_type_Reserved4:
             break;
     }
     REALM_ASSERT_DEBUG(false);
@@ -193,9 +204,10 @@ void col_type_deleg(Op& op, ColumnType type)
 
 
 size_t ColumnBase::get_size_from_type_and_ref(ColumnType type, ref_type ref,
-                                              Allocator& alloc) noexcept
+                                              Allocator& alloc,
+                                              bool nullable) noexcept
 {
-    GetSizeFromRef op(ref, alloc);
+    GetSizeFromRef op(ref, alloc, nullable);
     col_type_deleg(op, type);
     return op.m_size;
 }