+++ test/test_table.cpp
@@ -7049,37 +7049,80 @@ TEST(Table_getVersionCounterAfterRowAccessor)
     int_fast64_t ver = t.get_version_counter();
     int_fast64_t newVer;
 
-#define _CHECK_VER_BUMP() \
-    newVer = t.get_version_counter();\
-    CHECK_GREATER(newVer, ver); \
-    ver = newVer;
+    auto check_ver_bump = [&]() {
+        newVer = t.get_version_counter();
+        CHECK_GREATER(newVer, ver);
+        ver = newVer;
+    };
 
     t.set_bool(col_bool, 0, true);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_int(col_int, 0, 42);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_string(col_string, 0, "foo");
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_float(col_float, 0, 0.42f);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_double(col_double, 0, 0.42);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_olddatetime(col_date, 0, 1234);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_binary(col_binary, 0, BinaryData("binary", 7));
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_timestamp(col_timestamp, 0, Timestamp(777, 888));
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 
     t.set_null(0, 0);
-    _CHECK_VER_BUMP();
+    check_ver_bump();
 }
+
+
+TEST(Table_ColumnSizeFromRef)
+{        
+    // This test a bug where get_size_from_type_and_ref() returned off-by-one on nullable integer columns. 
+    // It seems to be only invoked from Table::get_size_from_ref() which is fast static method that lets
+    // you find the size of a Table without having to create an instance of it. This seems to be only done
+    // on subtables, so the bug has not been triggered in public.
+
+    Group g;
+    TableRef t = g.add_table("table");
+    t->add_column(type_Int, "int", true);
+
+    auto check_column_sizes = [this, &t](size_t add) {
+        t->add_empty_row(add);
+        for (size_t i = 0; i < add; ++i) {
+            t->set_int(0, i, i);
+        }
+        using tf = _impl::TableFriend;
+        size_t actual_num_cols = t->get_column_count();
+        Spec& t_spec = tf::get_spec(*t);
+        for (size_t col_ndx = 0; col_ndx < actual_num_cols; ++col_ndx) {
+            ColumnType col_type = t_spec.get_column_type(col_ndx);
+            ColumnBase& base = tf::get_column(*t, col_ndx);
+            ref_type col_ref = base.get_ref();
+            bool nullable = (t_spec.get_column_attr(col_ndx) & col_attr_Nullable) == col_attr_Nullable;
+            size_t col_size = ColumnBase::get_size_from_type_and_ref(col_type, col_ref, base.get_alloc(), nullable);
+            CHECK_EQUAL(col_size, t->size());
+        }
+    };
+
+    // Test internal nodes
+    check_column_sizes(REALM_MAX_BPNODE_SIZE);
+
+    // clear will enforce arrays to change from "compact" form to "normal" form
+    t->clear();
+    check_column_sizes(REALM_MAX_BPNODE_SIZE);
+
+    // Try with more levels in the tree
+    check_column_sizes(10 * REALM_MAX_BPNODE_SIZE);
+}
+
 
 #endif // TEST_TABLE