+++ release_notes.md
@@ -2,7 +2,9 @@
 
 ### Bugfixes
 
-* Lorem ipsum.
+* Fixed various crashes when using subtables. The crash will occur when the first column
+  of the subtable if of type col_type_Timestamp or if it is of type Bool, Int or OldDateTime
+  while also being nullable. Caused by bad static get_size_from_ref() methods of columns.
 
 ### Breaking changes